ml_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-decouple==3.8
torch==2.1.0
torchvision==0.16.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
transformers==4.35.2
python-dateutil==2.8.2

